+++ src/app/modules/marketplace/components/module-detail/module-detail.component.ts
@@ -8,8 +8,11 @@ import { ToolsService } from '../../services/tools-service/tools.service';
 import { Location } from '@angular/common';
 import { MediaMatcher } from '@angular/cdk/layout';
 import { OscarInferenceService } from '@app/modules/inference/services/oscar-inference.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { OscarServiceRequest } from '@app/shared/interfaces/oscar-service.interface';
+import { AppConfigService } from '@app/core/services/app-config/app-config.service';
+import { timer } from 'rxjs';
+import { TranslateService } from '@ngx-translate/core';
+import { SnackbarService } from '@app/shared/services/snackbar/snackbar.service';
 
 @Component({
     selector: 'app-module-detail',
@@ -24,19 +27,19 @@ export class ModuleDetailComponent implements OnInit {
         private authService: AuthService,
         private route: ActivatedRoute,
         private breadcrumbService: BreadcrumbService,
+        public translateService: TranslateService,
         public location: Location,
         private router: Router,
-        private _snackBar: MatSnackBar,
+        private snackbarService: SnackbarService,
         private changeDetectorRef: ChangeDetectorRef,
         private media: MediaMatcher
     ) {
-        authService.userProfileSubject.subscribe((profile) => {
-            this.userProfile = profile;
-        });
         if (this.location.path().includes('tools')) {
             this.isTool = true;
         }
-
+        if (authService.isAuthenticated()) {
+            authService.loadUserProfile();
+        }
         this.mobileQuery = this.media.matchMedia('(max-width: 650px)');
         this._mobileQueryListener = () => changeDetectorRef.detectChanges();
         this.mobileQuery.addEventListener('change', this._mobileQueryListener);
@@ -45,6 +48,7 @@ export class ModuleDetailComponent implements OnInit {
     modulesList = [];
     module!: Module;
     userProfile?: UserProfile;
+    popupWindow: Window | undefined | null;
 
     isLoading = false;
     isInferenceModule = false;
@@ -52,18 +56,15 @@ export class ModuleDetailComponent implements OnInit {
 
     mobileQuery: MediaQueryList;
     private _mobileQueryListener: () => void;
-
-    isLoggedIn() {
-        return this.authService.isAuthenticated();
-    }
-
-    isAuthorized() {
-        return this.userProfile?.isAuthorized;
-    }
+    private stopPolling$ = timer(180000);
 
     ngOnInit(): void {
         this.route.params.subscribe((params) => {
             this.isLoading = true;
+            this.authService.userProfileSubject.subscribe((profile) => {
+                this.userProfile = profile;
+            });
+
             if (this.isTool) {
                 this.toolsService.getTool(params['id']).subscribe((tool) => {
                     this.isLoading = false;
@@ -84,6 +85,14 @@ export class ModuleDetailComponent implements OnInit {
         });
     }
 
+    isLoggedIn() {
+        return this.authService.isAuthenticated();
+    }
+
+    isAuthorized() {
+        return this.userProfile?.isAuthorized;
+    }
+
     createOscarService() {
         this.isLoading = true;
         const requestBody: OscarServiceRequest = {
@@ -102,24 +111,14 @@ export class ModuleDetailComponent implements OnInit {
                         .navigate(['/inference'])
                         .then((navigated: boolean) => {
                             if (navigated) {
-                                this._snackBar.open(
+                                this.snackbarService.openSuccess(
                                     'Oscar service created with name ' +
-                                        serviceName,
-                                    'X',
-                                    {
-                                        duration: 3000,
-                                        panelClass: ['success-snackbar'],
-                                    }
+                                        serviceName
                                 );
                             } else {
-                                this._snackBar.open(
+                                this.snackbarService.openError(
                                     'Error while creating service with name ' +
-                                        serviceName,
-                                    'X',
-                                    {
-                                        duration: 3000,
-                                        panelClass: ['red-snackbar'],
-                                    }
+                                        serviceName
                                 );
                             }
                         });
@@ -129,5 +128,10 @@ export class ModuleDetailComponent implements OnInit {
                 this.isLoading = false;
             },
         });
+    }
+
+    createGradioDeployment() {
+        sessionStorage.setItem('moduleData', JSON.stringify(this.module));
+        window.open(`${window.location.href}/try-me-nomad`);
     }
 }