+++ src/app/modules/marketplace/components/module-train/datasets/dataset-creation-detail-component/dataset-creation-detail.component.ts
@@ -19,7 +19,6 @@ import {
     MatDialog,
     MatDialogRef,
 } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTabChangeEvent } from '@angular/material/tabs';
 import { AppConfigService } from '@app/core/services/app-config/app-config.service';
 import { ZenodoService } from '@app/modules/marketplace/services/zenodo-service/zenodo.service';
@@ -29,6 +28,7 @@ import {
     ZenodoDatasetVersion,
     ZenodoSimpleDataset,
 } from '@app/shared/interfaces/dataset.interface';
+import { SnackbarService } from '@app/shared/services/snackbar/snackbar.service';
 import { Observable, map, startWith } from 'rxjs';
 
 export function doiValidator(): ValidatorFn {
@@ -53,7 +53,7 @@ export class DatasetCreationDetailComponent implements OnInit {
         private changeDetectorRef: ChangeDetectorRef,
         private media: MediaMatcher,
         private fb: FormBuilder,
-        private _snackBar: MatSnackBar,
+        private snackbarService: SnackbarService,
         @Optional() @Inject(MAT_DIALOG_DATA) public data: ZenodoSimpleDataset
     ) {
         this.mobileQuery = this.media.matchMedia('(max-width: 650px)');
@@ -188,10 +188,7 @@ export class DatasetCreationDetailComponent implements OnInit {
             },
             error: () => {
                 this.zenodoFormGroup.get('zenodoDatasetSelect')?.disable();
-                this._snackBar.open('Error retrieving datasets', 'X', {
-                    duration: 3000,
-                    panelClass: ['red-snackbar'],
-                });
+                this.snackbarService.openError('Error retrieving datasets');
             },
         });
     }
@@ -269,10 +266,9 @@ export class DatasetCreationDetailComponent implements OnInit {
             },
             error: () => {
                 this.zenodoFormGroup.get('zenodoVersionSelect')?.disable();
-                this._snackBar.open('Error retrieving dataset versions', 'X', {
-                    duration: 3000,
-                    panelClass: ['red-snackbar'],
-                });
+                this.snackbarService.openError(
+                    'Error retrieving dataset versions'
+                );
             },
         });
     }
@@ -313,13 +309,8 @@ export class DatasetCreationDetailComponent implements OnInit {
                 error: () => {
                     this.zenodoFormGroup.get('zenodoVersionSelect')?.disable();
                     this.dialogLoading = false;
-                    this._snackBar.open(
-                        'Error retrieving dataset information',
-                        'X',
-                        {
-                            duration: 3000,
-                            panelClass: ['red-snackbar'],
-                        }
+                    this.snackbarService.openError(
+                        'Error retrieving dataset information'
                     );
                 },
             });